hide constructors of IDs
https://github.com/iand675/hs-opentelemetry/pull/49#discussion_r1070138527
